REFACTOR: Moved Art Fan Identity to proper package
Moved Art Fan Identity to proper package and updated reference
