+++ ART/android/sub_app/fermat-art-android-sub-app-fan-identity-bitdubai/src/main/java/com/bitdubai/sub_app/art_fan_identity/sessions/ArtFanUserIdentitySubAppSession.java
@@ -1,4 +1,4 @@
-package sessions;
+package com.bitdubai.sub_app.art_fan_identity.sessions;
 
 import com.bitdubai.fermat_android_api.layer.definition.wallet.abstracts.AbstractFermatSession;
 import com.bitdubai.fermat_api.layer.dmp_module.sub_app_manager.InstalledSubApp;
