+++ ART/android/sub_app/fermat-art-android-sub-app-fan-identity-bitdubai/src/main/java/com/bitdubai/sub_app/art_fan_identity/sessions/SessionConstants.java
@@ -1,4 +1,4 @@
-package sessions;
+package com.bitdubai.sub_app.art_fan_identity.sessions;
 
 /**
  * Created by Manuel Perez (darkpriestrelative@gmail.com) on 08/04/16.
