+++ ART/android/sub_app/fermat-art-android-sub-app-fan-identity-bitdubai/src/main/java/com/bitdubai/sub_app/art_fan_identity/util/CommonLogger.java
@@ -1,4 +1,4 @@
-package util;
+package com.bitdubai.sub_app.art_fan_identity.util;
 
 import android.util.Log;
